backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
supabase
fal-client
pydantic
httpx
pillow